Moving to sector_generators and updating code
