+++ geoips/plugins/modules/sector_generators/clat_clon_resolution_shape.py
